Added the 4th test case with input=5, output=Buzz
Why not 4? Well you can add 4. No one gonna blame you doing that. But
what difference of 4 from 1 and 2? Remember, test is also code we need
to maintain. Less codes to maintain is always better than more.

As a rule of thumb, add minimum test that "enough" for you to be
confident that the function is working properly.