Implement Dependency Injection using setter method
Create a new setter method setFizzBuzzRules to set our class attribute
"rules". Take dependencies (FizzRule02 and BuzzRule02) out of the class
and put it in the method getNewFizzBuzz in the test class. This way we
do "Inversion of Control" i.e. we control behavior of the class
FizzBuzz02 by injecting dependencies to it. (Configurable class)

Don't forget to run the test for the green bar.