Create new TransferRule03 class and replace it with FizzBuzzRule03
Of course, you can just change the returned string from "FizzBuzz" to
"Transfer" in the class FizzBuzzRule03. But you may need to rename the
class for readability. And what if the requirement is changed again by
going back to FizzBuzz? 

I think it is better idea to plug the old one out and plug the new one
in.